changezzz + start working on rest methods

+++ src/Discord.Net.Core/Entities/Messages/Polls/PollAnswer.cs
@@ -8,14 +8,14 @@ public readonly struct PollAnswer
     /// <summary>
     ///     Gets the Id of the answer.
     /// </summary>
-    public readonly ulong AnswerId;
+    public readonly uint AnswerId;
 
     /// <summary>
     ///     Gets the poll media of this answer.
     /// </summary>
     public readonly PollMedia PollMedia;
 
-    internal PollAnswer(ulong id, PollMedia media)
+    internal PollAnswer(uint id, PollMedia media)
     {
         AnswerId = id;
         PollMedia = media;
