+++ src/Discord.Net.Rest/API/Common/PollAnswer.cs
@@ -5,7 +5,7 @@ namespace Discord.API;
 internal class PollAnswer
 {
     [JsonProperty("answer_id")]
-    public ulong AnswerId { get; set; }
+    public uint AnswerId { get; set; }
 
     [JsonProperty("poll_media")]
     public PollMedia PollMedia { get; set; }
